+++ scripts/gameplay/GameplayContext.lua
@@ -16,6 +16,15 @@ function GameplayContext.new()
 		alertTimers = { -1.5, -1.5 },
 		bpmData = nil,
 		buttonDelta = 0,
+		buttonRatings = {
+			errorEarly = 0,
+			nearEarly = 0,
+			criticalEarly = 0,
+			sCritical = 0,
+			criticalLate = 0,
+			nearLate = 0,
+			errorLate = 0,
+		},
 		chain = 0,
 		chainTimer = 0,
 		critHit = false,
@@ -59,10 +68,22 @@ function GameplayContext:handleButton(delta, rating)
 		self.maxExScore = self.maxExScore + 5
 	end
 
-	if rating == 1 then
+	if rating == 0 then
+		if delta < 0 then
+			self.buttonRatings.errorEarly = self.buttonRatings.errorEarly + 1
+		else
+			self.buttonRatings.errorLate = self.buttonRatings.errorLate + 1
+		end
+	elseif rating == 1 then
 		self.buttonDelta = delta
 		self.critHit = false
 		self.exScore = self.exScore + 2
+
+		if delta < 0 then
+			self.buttonRatings.nearEarly = self.buttonRatings.nearEarly + 1
+		else
+			self.buttonRatings.nearLate = self.buttonRatings.nearLate + 1
+		end
 	elseif rating == 2 then
 		local absDelta = abs(delta)
 
@@ -74,8 +95,18 @@ function GameplayContext:handleButton(delta, rating)
 
 		if absDelta <= self.sCritWindow then
 			self.exScore = self.exScore + 5
+
+			if delta < 0 then
+				self.buttonRatings.sCritical = self.buttonRatings.sCritical + 1
+			end
 		else
 			self.exScore = self.exScore + 4
+
+			if delta < 0 then
+				self.buttonRatings.criticalEarly = self.buttonRatings.criticalEarly + 1
+			else
+				self.buttonRatings.criticalLate = self.buttonRatings.criticalLate + 1
+			end
 		end
 	end
 end
@@ -121,6 +152,16 @@ function GameplayContext:updateLaserAlerts(isRight)
 	elseif self.alertTimers[1] < -1 then
 		self.alertTimers[1] = 1
 	end
+end
+
+function GameplayContext:resetButtonRatings()
+	self.buttonRatings.errorEarly = 0
+	self.buttonRatings.nearEarly = 0
+	self.buttonRatings.criticalEarly = 0
+	self.buttonRatings.sCritical = 0
+	self.buttonRatings.criticalLate = 0
+	self.buttonRatings.nearLate = 0
+	self.buttonRatings.errorLate = 0
 end
 
 return GameplayContext